Added correct exports of files
