+++ lib/index.ts
@@ -1,4 +1,8 @@
-import { MegaMenuLayout } from "./components";
 import "./index.css";
 
-export { MegaMenuLayout };
+export * from "./components/MegaMenuLayout";
+export type { LinkComponentType } from "./hooks/useLinkComponent";
+
+export * from "./types/country";
+export * from "./types/icon";
+export * from "./types/megamenu";
